rental rendering correctly after video selected for rental
