+++ src/App.js
@@ -22,6 +22,10 @@ function App() {
 
   const [searchResults, setSearchResults] = useState([]);
 
+  const [selectedVideoTitle, setSelectedVideoTitle] = useState('');
+  const [selectedCustomerId, setSelectedCustomerId] = useState('');
+  const [selectedCustomerName, setSelectedCustomerName] = useState('');
+
   useEffect(() => {
     axios.get(BASE_URL + '/videos')
     .then((response) => {
@@ -77,11 +81,63 @@ function App() {
     })
   }
 
+  //load rental selections (ie if page refreshes??)
+  useEffect(() => {
+    const storedVideoTitle = sessionStorage.getItem('selectedVideoTitle');
+    const storedCustomerId = sessionStorage.getItem('selectedCustomerId');
+    const storedCustomerName = sessionStorage.getItem('selectedCustomerName');
+
+    setSelectedVideoTitle(storedVideoTitle);
+    setSelectedCustomerId(storedCustomerId);
+    setSelectedCustomerName(storedCustomerName);
+  }, [])
+
+  //rental checkout callback
+  const onRentalRequestCallback = (selectedCustomerId, selectedVideoTitle) => {
+    const rentalObject = {
+      // eslint-disable-next-line camelcase
+      customer_id: selectedCustomerId,
+      title: selectedVideoTitle,
+    }
+    axios.post(BASE_URL + '/rentals/'+ selectedVideoTitle + '/check-out', rentalObject )
+    .then((response) => {
+
+      setSelectedVideoTitle('');
+      setSelectedCustomerId('');
+      setSelectedCustomerName('');
+      sessionStorage.clear();
+
+      const message =`Checkout Complete!`; 
+      console.log(message)
+    })
+    .catch((error) => {
+      const message=`There was an error with your rental request. ${error.message}.`;
+      setErrorMessage(message);
+      console.log(message);
+    })
+  }
+
+  //video selection callback
+  const onSelectVideoForRentalCallback = (videoTitle) => {
+    sessionStorage.setItem('selectedVideoTitle', videoTitle)
+    setSelectedVideoTitle(videoTitle)
+    console.log(`${videoTitle} selected for rental`)
+  }
+
+  //user selection callback
+  const onSelectCustomerForRental = (customerId, customerName) => {
+    sessionStorage.setItem('selectedCustomerId', customerId)
+    sessionStorage.setItem('selectedCustomerName', customerName)
+    console.log(`${customerName}, id: ${customerId} selected for rental`)
+  }
+
+
   return (
     <Router>
       <div className="App">
         <Nav />
-        <Rental />
+        <Rental onRentalRequest={onRentalRequestCallback} selectedVideoTitle={selectedVideoTitle} selectedCustomerId={selectedCustomerId} selectedCustomerName={selectedCustomerName} />
+
         { errorMessage ? <h3 className='error-message'>{errorMessage}</h3> : 
           <Switch>
             <Route path="/" exact component={Home}/>
@@ -89,7 +145,9 @@ function App() {
             <Route path="/customerlist">
               <CustomerList customerList={customerList} baseURL={BASE_URL}/>
             </Route>
-            <Route path="/videos/:title" component={Video}/>
+            <Route path="/videos/:title">
+              <Video baseURL={BASE_URL}  onSelectVideoForRental={onSelectVideoForRentalCallback}/>
+            </Route>
             <Route path="/videolibrary">
               <VideoLibrary videoLibrary={videoLibrary} baseURL={BASE_URL} />
             </Route>
