+++ src/components/Rental.js
@@ -1,49 +1,8 @@
 import React, {useState, useEffect} from 'react';
 import axios from 'axios';
 
-const Rental = () => {
-  const [errorMessage, setErrorMessage] = useState('');
-  const [selectedVideoTitle, setSelectedVideoTitle] = useState('');
-  const [selectedCustomerId, setSelectedCustomerId] = useState('');
-  const [selectedCustomerName, setSelectedCustomerName] = useState('');
-
-  const storedVideoTitle = sessionStorage.getItem('selectedVideoTitle');
-  const storedCustomerId = sessionStorage.getItem('selectedCustomerId');
-  const storedCustomerName = sessionStorage.getItem('selectedCustomerName');
-
-  useEffect(() => {
-
-    setSelectedVideoTitle(storedVideoTitle);
-    setSelectedCustomerId(storedCustomerId);
-    setSelectedCustomerName(storedCustomerName);
-
-  }, [storedVideoTitle, storedCustomerId, storedCustomerName])
-
-  const onRentalRequest = () => {
-    const rentalObject = {
-      // eslint-disable-next-line camelcase
-      customer_id: selectedCustomerId,
-      title: selectedVideoTitle,
-    }
-
-    axios.post('http://localhost:3000/rentals/'+ selectedVideoTitle + '/check-out', rentalObject )
-    .then((response) => {
-
-      setSelectedVideoTitle('');
-      setSelectedCustomerId('');
-      setSelectedCustomerName('');
-      sessionStorage.clear();
-
-      const message =`Checkout Complete!`; 
-      console.log(message)
-    })
-    .catch((error) => {
-      const message=`There was an error with your rental request. ${error.message}.`;
-      setErrorMessage(message);
-      console.log(message);
-    })
-  }
-
+// const Rental = ({onRentalRequest, selectedCustomerId, selectedCustomerName, selectedVideoTitle}) => {
+const Rental = ({onRentalRequest, selectedCustomerName, selectedVideoTitle, selectedCustomerId}) => {
   return (
     <div>
       { (selectedVideoTitle) ? <div className='router__div--selected-item'> Selected Video for Rental: {selectedVideoTitle} </div> : null }
